refactor: seperated StixBuilder to logical elements
