+++ internal-enrichment/censys-enrichmentapis/.gitignore
@@ -15,4 +15,4 @@ tests/__pycache__
 # configuration file
 config.yml
 src/config.yml
-
+Taskfile.yml
