+++ internal-enrichment/censys-enrichmentapis/src/requirements.txt
@@ -3,3 +3,4 @@ pydantic~=2.13.5
 requests~=2.33.0
 validators==0.35.0
 connectors-sdk @ git+https://github.com/OpenCTI-Platform/connectors.git@master#subdirectory=connectors-sdk
+censys-platform~=0.16.1
